genrand_util: fix possibly signed char modulo
generate_random_str_list_buf() fills the buffer with random bytes
using generate_secret_buffer(), but later uses the values as
char (possibly signed) when calculating the index into the
character list.

On platforms where char is signed, this may result in a negative
index and out-of-bounds read.

Cast buffer values to uint8_t before calculating modulo to ensure
the index is non-negative.

Pair-Programmed-With: Dmitry Mikhalchenko <tascad@altlinux.org>

+++ lib/util/genrand_util.c
@@ -231,7 +231,7 @@ _PUBLIC_ char *generate_random_str_list_buf(char *buf,
 	generate_secret_buffer((uint8_t *)buf, len);
 
 	for (i=0; i<len; i++) {
-		buf[i] = list[buf[i] % list_len];
+		buf[i] = list[(uint8_t)buf[i] % list_len];
 	}
 
 	return buf;
